requirements.txt
pgmpy==0.1.25

openai==1.78.0
anthropic==0.51.0
google-generativeai==0.8.5
replicate==1.0.6

langchain==0.3.25
langchain-community==0.3.23
langchain-openai==0.3.16
langchain-anthropic==0.3.13
langchain-google-genai==2.0.10
langchain-deepseek==0.1.3

pandas==2.2.3
matplotlib==3.10.3
seaborn==0.13.2
tqdm==4.67.1
tabulate==0.9.0
PyPDF2==3.0.1
python-dotenv==1.1.0

xlsxwriter
rpy2